+++ RELEASES.md
@@ -5,6 +5,7 @@
 * Added `FromIterator` implementations for iterators of borrowed items.
 * Added `EnumSet::bit_index` and `EnumSet::is_bit_valid` methods to query the
   bit associated with an enum variant, or if a bit index is valid.
+* `derive(EnumSetType)` now uses `::core::assert!` instead of whichever `assert!` is visible.
 
 # Version 1.1.8 (2025-08-16)