+++ RELEASES.md
@@ -1,7 +1,7 @@
-# Unreleased
+# Version 1.1.9 (2025-08-17)
 
-* `derive(EnumSetType)` now uses `::core::assert!` instead of whichever
-  `assert!` is visible. (thanks @bugadani)
+* Fixes a bug that causes compilation failures when a crate (such as `defmt`)
+  exposes a macro called `assert!`. (thanks @bugadani)
 
 # Version 1.1.8 (2025-08-16)